+++ xlsx.go
@@ -261,4 +261,3 @@ func ReadXLSXSheetNames(filename string) ([]string, error) {
 
 	return f.GetSheetList(), nil
 }
-